+++ src/lib/runner_test.rs
@@ -1060,6 +1060,7 @@ fn create_watch_task_with_makefile_and_empty_object_options() {
         postpone: None,
         ignore_pattern: None,
         no_git_ignore: None,
+        why: None,
         watch: None,
     };
 
@@ -1111,6 +1112,7 @@ fn create_watch_task_with_makefile_and_all_object_options_and_cli_args() {
         postpone: Some(true),
         ignore_pattern: Some("tools/*".to_string()),
         no_git_ignore: Some(true),
+        why: Some(true),
         watch: Some(vec!["dir1".to_string(), "dir2".to_string()]),
     };
 
@@ -1146,19 +1148,20 @@ fn create_watch_task_with_makefile_and_all_object_options_and_cli_args() {
     make_command_line.push_str(" some_task 1 2 \"3 4\"");
 
     let args = task.args.unwrap();
-    assert_eq!(args.len(), 12);
+    assert_eq!(args.len(), 13);
     assert_eq!(args[0], "watch".to_string());
     assert_eq!(args[1], "-q".to_string());
     assert_eq!(args[2], "--postpone".to_string());
     assert_eq!(args[3], "-i".to_string());
     assert_eq!(args[4], "tools/*".to_string());
     assert_eq!(args[5], "--no-gitignore".to_string());
-    assert_eq!(args[6], "-w".to_string());
-    assert_eq!(args[7], "dir1".to_string());
-    assert_eq!(args[8], "-w".to_string());
-    assert_eq!(args[9], "dir2".to_string());
-    assert_eq!(args[10], "-x".to_string());
-    assert_eq!(args[11], make_command_line.to_string());
+    assert_eq!(args[6], "--why".to_string());
+    assert_eq!(args[7], "-w".to_string());
+    assert_eq!(args[8], "dir1".to_string());
+    assert_eq!(args[9], "-w".to_string());
+    assert_eq!(args[10], "dir2".to_string());
+    assert_eq!(args[11], "-x".to_string());
+    assert_eq!(args[12], make_command_line.to_string());
 }
 
 #[test]
@@ -1173,6 +1176,7 @@ fn create_watch_task_with_makefile_and_all_object_options() {
         postpone: Some(true),
         ignore_pattern: Some("tools/*".to_string()),
         no_git_ignore: Some(true),
+        why: Some(true),
         watch: Some(vec!["dir1".to_string(), "dir2".to_string()]),
     };
 
@@ -1207,19 +1211,20 @@ fn create_watch_task_with_makefile_and_all_object_options() {
     make_command_line.push_str(" some_task");
 
     let args = task.args.unwrap();
-    assert_eq!(args.len(), 12);
+    assert_eq!(args.len(), 13);
     assert_eq!(args[0], "watch".to_string());
     assert_eq!(args[1], "-q".to_string());
     assert_eq!(args[2], "--postpone".to_string());
     assert_eq!(args[3], "-i".to_string());
     assert_eq!(args[4], "tools/*".to_string());
     assert_eq!(args[5], "--no-gitignore".to_string());
-    assert_eq!(args[6], "-w".to_string());
-    assert_eq!(args[7], "dir1".to_string());
-    assert_eq!(args[8], "-w".to_string());
-    assert_eq!(args[9], "dir2".to_string());
-    assert_eq!(args[10], "-x".to_string());
-    assert_eq!(args[11], make_command_line.to_string());
+    assert_eq!(args[6], "--why".to_string());
+    assert_eq!(args[7], "-w".to_string());
+    assert_eq!(args[8], "dir1".to_string());
+    assert_eq!(args[9], "-w".to_string());
+    assert_eq!(args[10], "dir2".to_string());
+    assert_eq!(args[11], "-x".to_string());
+    assert_eq!(args[12], make_command_line.to_string());
 }
 
 #[test]
@@ -1234,6 +1239,7 @@ fn create_watch_task_with_makefile_and_false_object_options() {
         postpone: Some(false),
         ignore_pattern: None,
         no_git_ignore: Some(false),
+        why: Some(false),
         watch: None,
     };