+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Enhancement: New precompiled target: x86_64-unknown-linux-gnu #853 (thanks @steverusso)
 * Enhancement: New log level 'off' and cli flag --slient #846
+* Enhancement: Upgrade cargo-watch to 8.4.0 and add new 'why' watch option #849
 
 ### v0.36.7 (2023-04-23)